Systemd Arch Linux support, new features, and fixes

+++ Screenshot_LLM.sh
@@ -1,29 +1,45 @@
 #!/bin/bash
-
+i=
 # Define the lock file path
-LOCKFILE="/tmp/screen_llm.lock"
+LOCKFILE="/tmp/screenshot_llm.lock"
+
+# Define the virtual environment path
+# Embed trailing slash to retain relative path capability
+VENVPATH="${VENVPATH:-$HOME/.screenshot_llm/}"
 
 # Function to pause the script
 pause_script() {
-    echo "An error occurred. Pausing the script. Press [Enter] to continue..."
-    rm "$LOCKFILE"
-    read
+    # Determine if executed via systemd unit
+    if [ -z "$SCRLLM_SYSTEMD_UNIT" ]; then
+        echo "An error occurred. Pausing the script. Press [Enter] to continue..."
+        rm "$LOCKFILE"
+        read
+    fi
     exit 1
 }
 
+# Prepare env file
+if [[ -n "$SCRLLM_ENV_FILE" && ! -e "$SCRLLM_ENV_FILE" ]]; then
+    [ ! -e "${SCRLLM_ENV_FILE%/*}" ] && mkdir -p "${SCRLLM_ENV_FILE%/*}"
+    touch "$SCRLLM_ENV_FILE"
+fi
+
 # Check if the lock file exists
-if [ -e "$LOCKFILE" ]; then
-    echo "Another instance of the script is already running. Exiting."
-    pause_script
-else
-    # Create the lock file
-    touch "$LOCKFILE"
+if [ -z "$SCRLLM_SYSTEMD_UNIT" ]; then
+    if [ -e "$LOCKFILE" ]; then
+        echo "Another instance of the script is already running. Exiting."
+        pause_script
+    else
+        # Create the lock file
+        touch "$LOCKFILE"
+    fi
 fi
 
 # Ensure the virtual environment is created and activated
-if [ ! -d "venv" ]; then
+if [ ! -d "${VENVPATH}venv" ]; then
     echo "Virtual environment not found. Creating one..."
-    python3 -m venv venv
+    mkdir -p $VENVPATH
+    python -m venv "${VENVPATH}venv"
     if [ $? -ne 0 ]; then
         echo "Failed to create virtual environment."
         pause_script
@@ -31,7 +47,7 @@ if [ ! -d "venv" ]; then
 fi
 
 # Activate the virtual environment
-source venv/bin/activate
+source ${VENVPATH}venv/bin/activate
 
 # Check if the virtual environment was activated successfully
 if [ $? -ne 0 ]; then
@@ -41,7 +57,12 @@ fi
 
 # Install required Python packages if requirements.txt exists
 if [ -f "requirements.txt" ]; then
-    pip install -r requirements.txt > /dev/null 2>&1
+    if [ -z "$SCRLLM_SYSTEMD_UNIT" ]; then
+        pip install -r requirements.txt > /dev/null 2>&1
+    else
+        # Log everything when executed via systemd
+        pip install -r requirements.txt
+    fi
     if [ $? -ne 0 ]; then
         echo "Failed to install required Python packages."
         pause_script
@@ -49,23 +70,32 @@ if [ -f "requirements.txt" ]; then
 fi
 
 # Check if the Python script is already running
-if pgrep -f "python main.py" > /dev/null; then
-    echo "Python script is already running. Exiting."
-    rm "$LOCKFILE"
-    exit 1
+if [ -z "$SCRLLM_SYSTEMD_UNIT" ]; then
+    if pgrep -f "python main.py --screenshot_llm" > /dev/null; then
+        echo "Python script is already running. Exiting."
+        rm "$LOCKFILE"
+        exit 1
+    fi
 fi
 
-# Run the main Python script in the background with nohup
-nohup python main.py > output.log 2>&1 &
-
-# Check if the script was started successfully
-if [ $? -eq 0 ]; then
-    echo "Python script started successfully. Check 'output.log' for output."
-    read -p "Press [Enter] to exit..."
+# Run the main Python script in the background with nohup when run without systemd
+# Send a bogus argument to match on to avoid conflicts managing the process
+if [ -z "$SCRLLM_SYSTEMD_UNIT" ]; then
+    nohup python main.py --screenshot_llm > ${VENVPATH}/output.log 2>&1 &
 else
-    echo "Failed to start the Python script."
-    pause_script
+    python main.py --screenshot_llm
 fi
 
-# The lock file will be removed automatically when the script exits
-trap "rm -f $LOCKFILE" EXIT
+# Check if the script was started successfully when run without systemd
+if [ -z "$SCRLLM_SYSTEMD_UNIT" ]; then
+    if [ $? -eq 0 ]; then
+        echo "Python script started successfully. Check 'output.log' for output."
+        read -p "Press [Enter] to exit..."
+    else
+        echo "Failed to start the Python script."
+        pause_script
+    fi
+
+    # The lock file will be removed automatically when the script exits
+    trap "rm -f $LOCKFILE" EXIT
+fi
